More accurate coordinate system

+++ birdo/src/birdo/utilities/enemy.java
@@ -78,10 +78,10 @@ public class enemy extends player {
 		}
 
 		if (type == "charge") {
-			int deltaX = p.x - x;
-			int deltaY = p.y - y;
+			double deltaX = p.x - x;
+			double deltaY = p.y - y;
 
-			double theta = Math.atan((double) deltaY / (double) deltaX);
+			double theta = Math.atan(deltaY / deltaX);
 
 			dx = -1 * (int) (6 * Math.cos(theta));
 			dy = -1 * (int) (6 * Math.sin(theta));
@@ -126,12 +126,12 @@ public class enemy extends player {
 		}
 		if (type == "homing") {
 			feather f = new feather(this.x, this.y, false);
-			int deltaX = p.x - x;
-			int deltaY = p.y - y;
+			double deltaX = p.x - x;
+			double deltaY = p.y - y;
 
-			double theta = Math.atan((double) deltaY / (double) deltaX);
-			f.dx = -1 * (int) (5 * Math.cos(theta));
-			f.dy = -1 * (int) (5 * Math.sin(theta));
+			double theta = Math.atan(deltaY / deltaX);
+			f.dx = -1 * (5 * Math.cos(theta));
+			f.dy = -1 * (5 * Math.sin(theta));
 
 			feathers.add(f);
 		}
@@ -142,11 +142,11 @@ public class enemy extends player {
 
 			f.dx = -5;
 
-			f1.dx = -1 * (int) (5 * Math.cos(15 * Math.PI / 180));
-			f1.dy = -1 * (int) (5 * Math.sin(15 * Math.PI / 180));
+			f1.dx = -1 * (5 * Math.cos(15 * Math.PI / 180));
+			f1.dy = -1 * (5 * Math.sin(15 * Math.PI / 180));
 
-			f2.dx = -1 * (int) (5 * Math.cos(345 * Math.PI / 180));
-			f2.dy = -1 * (int) (5 * Math.sin(345 * Math.PI / 180));
+			f2.dx = -1 * (5 * Math.cos(345 * Math.PI / 180));
+			f2.dy = -1 * (5 * Math.sin(345 * Math.PI / 180));
 
 			feathers.add(f);
 			feathers.add(f1);
@@ -167,17 +167,17 @@ public class enemy extends player {
 			f3.dx = 5;
 			f3.dy = 0;
 
-			f1.dx = -1 * (int) (-5 * Math.cos(Math.PI / 4));
-			f1.dy = -1 * (int) (-5 * Math.sin(Math.PI / 4));
+			f1.dx = -1 * (-5 * Math.cos(Math.PI / 4));
+			f1.dy = -1 * (-5 * Math.sin(Math.PI / 4));
 
-			f2.dx = -1 * (int) (-5 * Math.cos(7 * Math.PI / 4));
-			f2.dy = -1 * (int) (-5 * Math.sin(7 * Math.PI / 4));
+			f2.dx = -1 * (-5 * Math.cos(7 * Math.PI / 4));
+			f2.dy = -1 * (-5 * Math.sin(7 * Math.PI / 4));
 
-			f4.dx = (int) (-5 * Math.cos(Math.PI / 4));
-			f4.dy = (int) (-5 * Math.sin(Math.PI / 4));
+			f4.dx = (-5 * Math.cos(Math.PI / 4));
+			f4.dy = (-5 * Math.sin(Math.PI / 4));
 
-			f5.dx = (int) (-5 * Math.cos(7 * Math.PI / 4));
-			f5.dy = (int) (-5 * Math.sin(7 * Math.PI / 4));
+			f5.dx = (-5 * Math.cos(7 * Math.PI / 4));
+			f5.dy = (-5 * Math.sin(7 * Math.PI / 4));
 
 			f6.dx = 0;
 			f6.dy = -5;
@@ -199,17 +199,17 @@ public class enemy extends player {
 			feather f2 = new feather(this.x + this.w / 2 - 6, this.y + this.h / 2 - 4, false);
 			feather f3 = new feather(this.x + this.w / 2 - 6, this.y + this.h / 2 - 4, false);
 
-			f.dx = -1 * (int) (5 * Math.cos(shotmultiplier * Math.PI / 12));
-			f.dy = -1 * (int) (5 * Math.sin(shotmultiplier * Math.PI / 12));
+			f.dx = -1 * (5 * Math.cos(shotmultiplier * Math.PI / 12));
+			f.dy = -1 * (5 * Math.sin(shotmultiplier * Math.PI / 12));
 
-			f1.dx = -1 * (int) (5 * Math.cos((shotmultiplier * Math.PI / 12) + Math.PI / 2));
-			f1.dy = -1 * (int) (5 * Math.sin((shotmultiplier * Math.PI / 12) + Math.PI / 2));
+			f1.dx = -1 * (5 * Math.cos((shotmultiplier * Math.PI / 12) + Math.PI / 2));
+			f1.dy = -1 * (5 * Math.sin((shotmultiplier * Math.PI / 12) + Math.PI / 2));
 
-			f2.dx = -1 * (int) (5 * Math.cos((shotmultiplier * Math.PI / 12) + Math.PI));
-			f2.dy = -1 * (int) (5 * Math.sin((shotmultiplier * Math.PI / 12) + Math.PI));
+			f2.dx = -1 * (5 * Math.cos((shotmultiplier * Math.PI / 12) + Math.PI));
+			f2.dy = -1 * (5 * Math.sin((shotmultiplier * Math.PI / 12) + Math.PI));
 
-			f3.dx = -1 * (int) (5 * Math.cos((shotmultiplier * Math.PI / 12) + 3 * Math.PI / 2));
-			f3.dy = -1 * (int) (5 * Math.sin((shotmultiplier * Math.PI / 12) + 3 * Math.PI / 2));
+			f3.dx = -1 * (5 * Math.cos((shotmultiplier * Math.PI / 12) + 3 * Math.PI / 2));
+			f3.dy = -1 * (5 * Math.sin((shotmultiplier * Math.PI / 12) + 3 * Math.PI / 2));
 
 			shotmultiplier++;
 
@@ -226,18 +226,18 @@ public class enemy extends player {
 			feather f4 = new feather(this.x, this.y, false);
 			feather f5 = new feather(this.x, this.y, false);
 
-			f1.dx = -1 * (int) (5 * Math.cos(Math.PI / 8));
-			f1.dy = -1 * (int) (5 * Math.sin(Math.PI / 8));
-			f2.dx = -1 * (int) (5 * Math.cos(Math.PI / 4));
-			f2.dy = -1 * (int) (5 * Math.sin(Math.PI / 4));
-			f3.dx = -1 * (int) (5 * Math.cos(3 * Math.PI / 8));
-			f3.dy = -1 * (int) (5 * Math.sin(3 * Math.PI / 8));
-			f4.dx = (int) (5 * Math.cos(5 * Math.PI / 8));
-			f4.dy = (int) (5 * Math.sin(5 * Math.PI / 8));
-			f5.dx = (int) (5 * Math.cos(6 * Math.PI / 8));
-			f5.dy = (int) (5 * Math.sin(6 * Math.PI / 8));
-			f.dx = (int) (5 * Math.cos(7 * Math.PI / 8));
-			f.dy = (int) (5 * Math.sin(7 * Math.PI / 8));
+			f1.dx = -1 * (5 * Math.cos(Math.PI / 8));
+			f1.dy = -1 * (5 * Math.sin(Math.PI / 8));
+			f2.dx = -1 * (5 * Math.cos(Math.PI / 4));
+			f2.dy = -1 * (5 * Math.sin(Math.PI / 4));
+			f3.dx = -1 * (5 * Math.cos(3 * Math.PI / 8));
+			f3.dy = -1 * (5 * Math.sin(3 * Math.PI / 8));
+			f4.dx = (5 * Math.cos(5 * Math.PI / 8));
+			f4.dy = (5 * Math.sin(5 * Math.PI / 8));
+			f5.dx = (5 * Math.cos(6 * Math.PI / 8));
+			f5.dy = (5 * Math.sin(6 * Math.PI / 8));
+			f.dx = (5 * Math.cos(7 * Math.PI / 8));
+			f.dy = (5 * Math.sin(7 * Math.PI / 8));
 
 			feathers.add(f1);
 			feathers.add(f2);
